pyproject.toml
[project]
name = "yolo-code-assistant"
version = "0.1.0"
description = "RAG-based code assistant for Ultralytics YOLO"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0,<0.115.0",
    "uvicorn>=0.24.0,<0.32.0",
    "pymongo[srv]>=4.6.0,<5.0.0",
    "sentence-transformers>=2.2.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "requests>=2.31.0,<3.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "streamlit>=1.30.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<8.0.0",
    "ruff>=0.1.0,<1.0.0",
    "black>=23.0.0,<25.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]